+++ cognite/client/data_classes/geospatial.py
@@ -1,4 +1,4 @@
-from typing import Any, Dict
+from typing import Any, Dict, List, Union
 
 from cognite.client import utils
 from cognite.client.data_classes._base import CogniteResource, CogniteResourceList
@@ -40,7 +40,9 @@ class FeatureTypeList(CogniteResourceList):
 class PropertyAndSearchSpec:
     """A representation of a feature type property and search spec."""
 
-    def __init__(self, properties: Dict[str, Any] = None, search_spec: Dict[str, Any] = None):
+    def __init__(
+        self, properties: Union[Dict[str, Any], List[str]] = None, search_spec: Union[Dict[str, Any], List[str]] = None
+    ):
         self.properties = properties
         self.search_spec = search_spec
 
@@ -48,9 +50,16 @@ class PropertyAndSearchSpec:
 class FeatureTypeUpdate:
     """A representation of a feature type update in the geospatial api."""
 
-    def __init__(self, external_id: str = None, add: PropertyAndSearchSpec = None, cognite_client=None):
+    def __init__(
+        self,
+        external_id: str = None,
+        add: PropertyAndSearchSpec = None,
+        remove: PropertyAndSearchSpec = None,
+        cognite_client=None,
+    ):
         self.external_id = external_id
         self.add = add
+        self.remove = remove
         self._cognite_client = cognite_client
 
 
@@ -78,7 +87,44 @@ class Feature(CogniteResource):
 
 
 def _is_geometry_type(property_type: str):
-    return property_type in {"POINT", "POLYGON", "MULTIPOLYGON", "GEOMETRYCOLLECTION"}
+    return property_type in {
+        "GEOMETRY",
+        "POINT",
+        "LINESTRING",
+        "POLYGON",
+        "MULTIPOINT",
+        "MULTILINESTRING",
+        "MULTIPOLYGON",
+        "GEOMETRYCOLLECTION",
+        "GEOMETRYZ",
+        "POINTZ",
+        "LINESTRINGZ",
+        "POLYGONZ",
+        "MULTIPOINTZ",
+        "MULTILINESTRINGZ",
+        "MULTIPOLYGONZ",
+        "GEOMETRYCOLLECTIONZ",
+        "GEOMETRYM",
+        "POINTM",
+        "LINESTRINGM",
+        "POLYGONM",
+        "MULTIPOINTM",
+        "MULTILINESTRINGM",
+        "MULTIPOLYGONM",
+        "GEOMETRYCOLLECTIONM",
+        "GEOMETRYZM",
+        "POINTZM",
+        "LINESTRINGZM",
+        "POLYGONZM",
+        "MULTIPOINTZM",
+        "MULTILINESTRINGZM",
+        "MULTIPOLYGONZM",
+        "GEOMETRYCOLLECTIONZM",
+    }
+
+
+def _is_reserved_property(property_name: str):
+    return property_name.startswith("_") or property_name in {"externalId", "createdTime", "lastUpdatedTime"}
 
 
 class FeatureList(CogniteResourceList):
@@ -103,28 +149,55 @@ class FeatureList(CogniteResourceList):
         return gdf
 
     @staticmethod
-    def from_geopandas(feature_type: FeatureType, gdf: "geopandas.GeoDataFrame") -> "FeatureList":
+    def from_geopandas(
+        feature_type: FeatureType,
+        geodataframe: "geopandas.GeoDataFrame",
+        external_id_column: str = "externalId",
+        property_column_mapping: Dict[str, str] = None,
+    ) -> "FeatureList":
         """Convert a GeoDataFrame instance into a FeatureList.
 
         Args:
             feature_type (FeatureType): The feature type the features will conform to
-            gdf (GeoDataFrame): the geodataframe instance to convert into features
+            geodataframe (GeoDataFrame): the geodataframe instance to convert into features
+            external_id_column: the geodataframe column to use for the feature external id
+            property_column_mapping: provides a mapping from featuretype property names to geodataframe columns
 
         Returns:
             FeatureList: The list of features converted from the geodataframe rows.
+
+        Examples:
+
+            Create features from a geopandas dataframe:
+
+                >>> from cognite.client import CogniteClient
+                >>> c = CogniteClient()
+                >>> my_feature_type = ... # some feature type with 'position' and 'temperature' properties
+                >>> my_geodataframe = ...  # some geodataframe with 'center_xy', 'temp' and 'id' columns
+                >>> feature_list = FeatureList.from_geopandas(feature_type=my_feature_type, geodataframe=my_geodataframe,
+                >>>     external_id_column="id", property_column_mapping={'position': 'center_xy', 'temperature': 'temp'})
+                >>> created_features = c.geospatial.create_features(my_feature_type.external_id, feature_list)
+
         """
         features = []
-        for _, row in gdf.iterrows():
-            feature = Feature(external_id=row["externalId"])
-            for attr in feature_type.properties.items():
-                attr_name = attr[0]
-                attr_type = attr[1]["type"]
-                if attr_name.startswith("_"):
+        if property_column_mapping is None:
+            property_column_mapping = {prop_name: prop_name for (prop_name, _) in feature_type.properties.items()}
+        for _, row in geodataframe.iterrows():
+            feature = Feature(external_id=row[external_id_column])
+            for prop in feature_type.properties.items():
+                prop_name = prop[0]
+                prop_type = prop[1]["type"]
+                prop_optional = prop[1].get("optional", False)
+                if _is_reserved_property(prop_name):
                     continue
-                if _is_geometry_type(attr_type):
-                    setattr(feature, attr_name, {"wkt": row[attr_name].wkt})
+                column_name = property_column_mapping[prop[0]]
+                column_value = row[column_name]
+                if column_value is None and prop_optional:
+                    continue
+                if _is_geometry_type(prop_type):
+                    setattr(feature, prop_name, {"wkt": column_value.wkt})
                 else:
-                    setattr(feature, attr_name, row[attr_name])
+                    setattr(feature, prop_name, column_value)
             features.append(feature)
         return FeatureList(features)
 
@@ -132,9 +205,7 @@ class FeatureList(CogniteResourceList):
 class FeatureAggregate(CogniteResource):
     """A result of aggregating features in geospatial api."""
 
-    def __init__(self, cognite_client=None, **aggregates):
-        for key in aggregates:
-            setattr(self, key, aggregated[key])
+    def __init__(self, cognite_client=None):
         self._cognite_client = cognite_client
 
     @classmethod